Fix DeriveKeyPair definition
`slice` is defined with 3 arguments, but only two are passed. Clearly mention the start.
Strings are used as byte string, such as "msg" for instance. Define this notation.
`DerivePublicKey` takes a public key `(n, e)`. However, `e` is not used, nor it is passed to `DeriveKeyPair` which calls `DerivePublicKey`. Simply drop this argument.